pyproject.toml
[tool.poetry]
name = "telebot"
version = "0.1.0"
description = ""
authors = ["ren <denis30510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pytelegrambotapi = "^4.16.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
